Use inlinable strings for debug labels
This is a work in progress. This currently assumes valid utf-8
as the label input, but this can be relaxed once gfx-hal is
updated to accept `*const c_char` or `&CStr` for labels instead
of a rust `&str`. At that point the inlinable `Label` string
should probably be moved to gfx-hal as well.

This version of an inline string accepts up to 62 bytes
before spilling onto the heap. Most other options in the
rust ecosystem seem to limit to 30 bytes or less.